Extract generation of spend transaction
